runsc/cmd/sandboxsetup: export gofer setup helpers for reuse
Building a custom gVisor gofer that serves storage volumes over LisaFS
requires duplicating unexported setup helpers from runsc/cmd/ like
applyCaps, resolveSymlinks, adjustMountOptions, pivotRoot, copyFile,
prepareArgs, execProcUmounter, setCapsAndCallSelf, and intFlags.
These must be manually diffed and re-synced on every gVisor version bump.

The functions are already well-factored: they take explicit parameters,
carry no hidden state, and are unexported only because they live in a CLI
package rather than because they have invariants worth protecting. This
change adds exported equivalents in a new runsc/cmd/sandboxsetup/ package
(with public visibility) and updates all internal callers to use the new
package directly, leaving no stub wrappers behind.

Also fixes a pre-existing bug in CallSelfAsNobody where the error
messages for SYS_SETGID and SYS_SETUID were swapped.
