+++ runsc/cmd/sandboxsetup/BUILD
@@ -1,0 +1,29 @@
+load("//tools:defs.bzl", "go_library")
+
+package(
+    default_applicable_licenses = ["//:license"],
+    licenses = ["notice"],
+)
+
+go_library(
+    name = "sandboxsetup",
+    srcs = [
+        "caps.go",
+        "flags.go",
+        "fs.go",
+        "process.go",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/fd",
+        "//pkg/log",
+        "//runsc/cmd/util",
+        "//runsc/config",
+        "//runsc/flag",
+        "//runsc/specutils",
+        "//runsc/starttime",
+        "@com_github_moby_sys_capability//:go_default_library",
+        "@com_github_opencontainers_runtime_spec//specs-go:go_default_library",
+        "@org_golang_x_sys//unix:go_default_library",
+    ],
+)
